RocksDB Move common methods up to RocksBaseValue
