+++ versioned/tiered/rocksdb/src/main/java/com/dremio/nessie/versioned/store/rocksdb/RocksBaseValue.java
@@ -18,6 +18,7 @@ package com.dremio.nessie.versioned.store.rocksdb;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -90,5 +91,31 @@ abstract class RocksBaseValue<C extends BaseValue<C>> implements BaseValue<C> {
    */
   Entity toEntity(Stream<Id> idStream, int position) {
     return idStream.collect(Collectors.toList()).get(position).toEntity();
+  }
+
+  /**
+   * Evaluates if the stream of Id meets the Condition Function
+   * @param function the function
+   * @param stream
+   * @return
+   */
+  boolean evaluateStream(Function function, Stream<Id> stream) {
+    // stream is a list. EQUALS will either compare a specified position or the whole list.
+    List<String> path = Arrays.asList(function.getPath().split(Pattern.quote(".")));
+    String segment = path.get(0);
+    if (path.size() == 1) {
+      if (function.getOperator().equals(Function.EQUALS)) {
+        List<String> arguments = splitArrayString(segment);
+        if (arguments.size() == 1) { // compare complete list
+          return (toEntity(stream).equals(function.getValue()));
+        } else if (arguments.size() == 2) { // compare individual element of list
+          int position = Integer.parseInt(arguments.get(1));
+          return (toEntity(stream, position).equals(function.getValue()));
+        }
+      } else if (function.getOperator().equals(Function.SIZE)) {
+        return (stream.collect(Collectors.toList()).size() == (int)function.getValue().getNumber());
+      }
+    }
+    return false;
   }
 }
