+++ versioned/tiered/rocksdb/src/main/java/com/dremio/nessie/versioned/store/rocksdb/RocksL1.java
@@ -143,32 +143,6 @@ public class RocksL1 extends RocksBaseValue<L1> implements L1, Evaluator {
     return false;
   }
 
-  /**
-   * Evaluates if the stream of Id meets the Condition Function
-   * @param function the function
-   * @param stream
-   * @return
-   */
-  boolean evaluateStream(Function function, Stream<Id> stream) {
-    // stream is a list. EQUALS will either compare a specified position or the whole list.
-    List<String> path = Arrays.asList(function.getPath().split(Pattern.quote(".")));
-    String segment = path.get(0);
-    if (path.size() == 1) {
-      if (function.getOperator().equals(Function.EQUALS)) {
-        List<String> arguments = splitArrayString(segment);
-        if (arguments.size() == 1) { // compare complete list
-          return (toEntity(stream).equals(function.getValue()));
-        } else if (arguments.size() == 2) { // compare individual element of list
-          int position = Integer.parseInt(arguments.get(1));
-          return (toEntity(stream, position).equals(function.getValue()));
-        }
-      } else if (function.getOperator().equals(Function.SIZE)) {
-        return (tree.collect(Collectors.toList()).size() == (int)function.getValue().getNumber());
-      }
-    }
-    return false;
-  }
-
   public Stream<Id> getAncestors() {
     return parentList;
   }
